# src/main/java/com/tpastushok/cosmocats/dto/competitors/observer/CompetitorsObserverResponseDto.java
package com.tpastushok.cosmocats.dto.competitors.observer;

import lombok.Builder;
import lombok.Value;
import lombok.extern.jackson.Jacksonized;

import java.util.List;
import java.util.UUID;

@Value
@Builder
@Jacksonized
public class CompetitorsObserverResponseDto {
    UUID productUuid;
    List<CompetitorStorePrice> stores;
}